dev-requirements.txt
-r requirements.txt
pre-commit==2.16.0
pre-commit-hooks==4.1.0
black==21.12b0
flake8==4.0.1
pycln==1.1.0
isort==5.10.1
mypy==0.931
mypy-extensions==0.4.3
sqlalchemy-stubs==0.4
pytest
allure-pytest
